backend/python_app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.2.0
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
alembic==1.13.1
pandas==2.3.2
flake8==6.1.0
pytest==7.4.3
httpx==0.25.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.1
python-json-logger==2.0.7
requests==2.32.3
beautifulsoup4==4.12.3
fake-useragent==1.5.1
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
openpyxl==3.1.2